// src/constants/TokenTypes.d.ts
export declare enum TokenTypes {
    OTHER = "other",
    COLOR = "color",
    IMPLICIT = "implicit",
    BORDER_RADIUS = "borderRadius",
    SIZING = "sizing",
    SPACING = "spacing",
    TEXT = "text",
    TYPOGRAPHY = "typography",
    OPACITY = "opacity",
    BORDER_WIDTH = "borderWidth",
    BOX_SHADOW = "boxShadow",
    FONT_FAMILIES = "fontFamilies",
    FONT_WEIGHTS = "fontWeights",
    LINE_HEIGHTS = "lineHeights",
    FONT_SIZES = "fontSizes",
    LETTER_SPACING = "letterSpacing",
    PARAGRAPH_SPACING = "paragraphSpacing",
    TEXT_DECORATION = "textDecoration",
    TEXT_CASE = "textCase"
}